% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bbox_buffer.R
\name{add_bbox_buffer}
\alias{add_bbox_buffer}
\title{Add a uniform buffer around a bounding box}
\usage{
add_bbox_buffer(bbox, distance, distance_unit = "meters")
}
\arguments{
\item{bbox}{The original bounding box to add a buffer around. If not already
a [terrainr_bounding_box] object, will be converted.}

\item{distance}{The distance to add to the buffer.}

\item{distance_unit}{The units of the distance to add to the buffer, passed
to [convert_distance].}
}
\value{
A \code{\link{terrainr_bounding_box}} object.
}
\description{
This function calculates the great circle distance both corners of your
bounding box are from the centroid and extends those by a set distance. Due
to using great circle distance, calculations will not be exact.
}
\examples{
add_bbox_buffer(
  list(
    c(lat = 44.04905, lng = -74.01188),
    c(lat = 44.17609, lng = -73.83493)
  ),
  10
)
}
