% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bbox_buffer.R
\name{addbuff}
\alias{addbuff}
\alias{add_bbox_buffer}
\alias{set_bbox_side_length}
\title{Add a uniform buffer around a bounding box}
\usage{
add_bbox_buffer(bbox, distance, distance_unit = "meters", divisible = NULL)

set_bbox_side_length(bbox, distance, distance_unit = "meters")
}
\arguments{
\item{bbox}{The original bounding box to add a buffer around. If not already
a \code{\link{terrainr_bounding_box}} object, will be converted.}

\item{distance}{The distance to add to the buffer.}

\item{distance_unit}{The units of the distance to add to the buffer, passed
to \code{\link{convert_distance}}.}

\item{divisible}{Numeric: Extend the top right and bottom left corners so
that each side is divisible by \code{divisible}. Leave set to \code{NULL} to
not extend. This argument is in the same units as \code{distance_unit}.}
}
\value{
A [terrainr_bounding_box] object.
}
\description{
[add_bbox_buffer] calculates the great circle distance both corners of
your bounding box are from the centroid and extends those by a set distance.
Due to using Haversine "great circle" distance, calculations will not be
exact.

[set_bbox_side_length] is a thin wrapper around [add_bbox_buffer] which sets
all sides of the bounding box to (approximately) a specified length.
}
\examples{
add_bbox_buffer(
  list(
    c(lat = 44.04905, lng = -74.01188),
    c(lat = 44.17609, lng = -73.83493)
  ),
  10
)
set_bbox_side_length(
  list(
    c(lat = 44.04905, lng = -74.01188),
    c(lat = 44.17609, lng = -73.83493)
  ),
  4000
)
}
\seealso{
Other utilities: 
\code{\link{calc_haversine_distance}()},
\code{\link{convert_distance}()},
\code{\link{deg_to_rad}()},
\code{\link{get_bbox_centroid}()},
\code{\link{get_bbox}()},
\code{\link{point_from_distance}()},
\code{\link{rad_to_deg}()}
}
\concept{utilities}
