% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_haversine_distance.R
\name{calc_haversine_distance}
\alias{calc_haversine_distance}
\title{Extract latitude and longitude from a provided object}
\usage{
calc_haversine_distance(point_1, point_2)
}
\arguments{
\item{point_1, point_2}{Coordinate pairs (as length-2 numeric vectors with the
names "lat" and "lng") to calculate distance between.}
}
\value{
A vector of length 2 containing object latitude and longitude
}
\description{
This is an internal utility function to convert bounding boxes into
coordinate pairs.
}
\examples{
calc_haversine_distance(
  c(lat = 44.121268, lng = -73.903734),
  c(lat = 43.121268, lng = -74.903734)
)

}
\seealso{
Other utilities: 
\code{\link{add_bbox_buffer}()},
\code{\link{convert_distance}()},
\code{\link{deg_to_rad}()},
\code{\link{get_bbox_centroid}()},
\code{\link{get_coord_bbox}()},
\code{\link{point_from_distance}()},
\code{\link{rad_to_deg}()}
}
\concept{utilities}
\keyword{internal}
