% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_overlays.R
\name{combine_overlays}
\alias{combine_overlays}
\title{Combine multiple image overlays into a single file}
\usage{
combine_overlays(..., output_file = NULL, transparency = 0)
}
\arguments{
\item{...}{File paths for images to be combined. Note that combining TIFF
images requires the \code{tiff} package be installed.}

\item{output_file}{Optionally, the path to save the resulting image to. Can
be any format accepted by \link[magick:editing]{magick::image_read}.}

\item{transparency}{}
}
\value{
The combined image as a \code{magick-image} object, returned invisibly.
}
\description{
This function combines any number of images into a single file, which may
then be further processed as an image or transformed into an image overlay.
}
\examples{
\dontrun {

# Generate points and download orthoimagery
mt_elbert_points <- data.frame(
lat = runif(100, min = 39.11144, max = 39.12416),
lng = runif(100, min = -106.4534, max = -106.437)
)

mt_elbert_bbox <- get_coord_bbox(data = mt_elbert_points,
                                 lat = "lat",
                                 lng = "lng")

output_files <- get_tiles(bbox = mt_elbert_bbox,
                          output_prefix = tempfile(),
                          services = c("ortho"))

# Merge orthoimagery into a single file
ortho_merged <- merge_rasters(input_rasters = output_files[1],
                              output_raster = tempfile(fileext = ".tif"))

# Create an sf dataset from our points
mt_elbert_sf <- sf::st_as_sf(mt_elbert_points, coords = c("lng", "lat"))
sf::st_crs(mt_elbert_sf) <- sf::st_crs(4326)

# Convert our points into an overlay
mt_elbert_overlay <- vector_to_overlay(mt_elbert_sf,
                                       ortho_merged[[1]],
                                       size = 15,
                                       color = "red",
                                       na.rm = TRUE)

# Combine the overlay with our orthoimage
ortho_with_points <- combine_overlays(ortho_merged[[1]],
                                      mt_elbert_overlay)

}
}
\seealso{
Other data manipulation functions: 
\code{\link{georeference_overlay}()},
\code{\link{merge_rasters}()},
\code{\link{raster_to_raw_tiles}()},
\code{\link{vector_to_overlay}()}

Other overlay creation functions: 
\code{\link{georeference_overlay}()},
\code{\link{vector_to_overlay}()}
}
\concept{data manipulation functions}
\concept{overlay creation functions}
