% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert decimal degrees to radians}
\usage{
deg_to_rad(deg)
}
\arguments{
\item{deg}{A vector of values, in decimal degrees, to convert to radians}
}
\value{
A vector of the same length in radians
}
\description{
Convert decimal degrees to radians
}
\examples{
deg_to_rad(360)
rad_to_deg(deg_to_rad(360))
}
\seealso{
Other utilities: 
\code{\link{addbuff}},
\code{\link{calc_haversine_distance}()},
\code{\link{convert_distance}()},
\code{\link{get_bbox_centroid}()},
\code{\link{get_bbox}()},
\code{\link{point_from_distance}()},
\code{\link{rad_to_deg}()}
}
\concept{utilities}
