% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{export_bounding_box}
\alias{export_bounding_box}
\title{Convert a terrainr_bounding_box object to a base list}
\usage{
export_bounding_box(bbox)
}
\arguments{
\item{bbox}{A [terrainr::terrainr_bounding_box] object}
}
\value{
A list with coordinate pairs representing the lower left and upper
right corners of a bounding box
}
\description{
Convert a terrainr_bounding_box object to a base list
}
\examples{
# Create bounding box from coordinates:
bbox <- terrainr_bounding_box(bl = c(44.05003, -74.01164),
                              tr = c(44.17538, -73.83500))
export_bounding_box(bbox)

}
