% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{export_coord_pair}
\alias{export_coord_pair}
\title{Convert a terrainr_coordinate_pair object to a base vector.}
\usage{
export_coord_pair(coord_pair)
}
\arguments{
\item{coord_pair}{A \code{\link{terrainr_coordinate_pair}} object}
}
\value{
A vector with a coordinate pair in (latitude, longitude) format
}
\description{
Convert a terrainr_coordinate_pair object to a base vector.
}
\examples{
coord_pair <- terrainr_coordinate_pair(c(lat = 44.05003, lng = -74.01164))
export_coord_pair(coord_pair)
}
\seealso{
Other classes and related functions: 
\code{\link{terrainr_coordinate_pair-class}},
\code{\link{terrainr_coordinate_pair}}
}
\concept{classes and related functions}
