% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbox_centroid.R
\name{get_bbox_centroid}
\alias{get_bbox_centroid}
\title{Get the great-circle centroid for a given bounding box}
\usage{
get_bbox_centroid(bbox)
}
\arguments{
\item{bbox}{The bounding box to find a centroid for. If not already
a [terrainr::terrainr_bounding_box] object, will be converted.}
}
\description{
Get the great-circle centroid for a given bounding box
}
\examples{
get_bbox_centroid(
  list(
    c(lat = 44.04905, lng = -74.01188),
    c(lat = 44.17609, lng = -73.83493)
  )
)
}
