% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_centroid}
\alias{get_centroid}
\title{Get the great-circle centroid for latitude/longitude data}
\usage{
get_centroid(lat, lng)
}
\arguments{
\item{lat}{A vector of latitudes in degrees.}

\item{lng}{A vector of longitudes in degrees.}
}
\value{
A latitude/longitude
}
\description{
Get the great-circle centroid for latitude/longitude data
}
\seealso{
Other utilities: 
\code{\link{addbuff}},
\code{\link{calc_haversine_distance}()},
\code{\link{deg_to_rad}()},
\code{\link{rad_to_deg}()}
}
\concept{utilities}
\keyword{internal}
