% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heightmap.R
\name{get_heightmap}
\alias{get_heightmap}
\title{Get USGS Digital Elevation Model heightmap for an area}
\usage{
get_heightmap(bbox, output_file, verbose = FALSE)
}
\arguments{
\item{bbox}{A bounding box representing the lower left and upper right corner
of the area to retrieve a heightmap for. If not a
[terrainr::terrainr_bounding_box] object, it will be coerced to one.}

\item{output_file}{The file path to save the merged output raster to. Must
be a GeoTIFF file.}

\item{verbose}{Logical: should [terrainr::hit_heightmap_api] run in verbose
mode (printing out the number of tries it takes to retrieve each tile)?}
}
\description{
This function splits the area contained within a bounding box into a set of
tiles, retrieves USGS 3DEP heightmaps for each tile, and then merges those
tiles together into a final output raster.
}
\examples{
\dontrun{
simulated_data <- data.frame(
  id = seq(1, 100, 1),
  lat = runif(100, 44.04905, 44.17609),
  lng = runif(100, -74.01188, -73.83493)
)

bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
bbox <- add_bbox_buffer(bbox, 100)
bbox <- make_unity_friendly(bbox)
temptiff <- tempfile(fileext = ".tif")
get_heightmap_tiles(temptiff)
raster_to_raw_tiles(temptiff, tempfile())
}

}
