% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiles.R
\name{get_tiles}
\alias{get_tiles}
\title{Get USGS elevation and imagery tiles for an area}
\usage{
get_tiles(
  bbox,
  output_prefix,
  side_length,
  extend = FALSE,
  elevation = TRUE,
  ortho = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{bbox}{A bounding box representing the lower left and upper right corner
of the area to retrieve a heightmap for. If not a
[terrainr::terrainr_bounding_box] object, it will be coerced to one.}

\item{output_prefix}{The file prefix to use when saving tiles.}

\item{side_length}{The length, in meters (and therefore pixels), of each side
of tiles to download.}

\item{extend}{Logical: if the side length of the bounding box isn't divisible
by \code{side_length}, should it be extended to be divisible?}

\item{elevation}{Logical: download elevation tiles for the}

\item{verbose}{Logical: should tile retrieval functions run in verbose mode?}
}
\value{
The original bounding box, invisibly.
}
\description{
This function splits the area contained within a bounding box into a set of
tiles, and retrieves USGS 3DEP heightmaps and NAIP orthoimagery for each tile.
Tiles are downloaded at a resolution of 1 meter per pixel.
}
\examples{
\dontrun{
simulated_data <- data.frame(
  id = seq(1, 100, 1),
  lat = runif(100, 44.04905, 44.17609),
  lng = runif(100, -74.01188, -73.83493)
)

bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
bbox <- add_bbox_buffer(bbox, 100)
bbox <- make_unity_friendly(bbox)
get_tiles(bbox, tempfile(), 4096)
}

}
