% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unity_friendly.R
\name{make_unity_friendly}
\alias{make_unity_friendly}
\title{Resize bounding box to reduce number of blank tiles created}
\usage{
make_unity_friendly(bbox, size = 4097)
}
\arguments{
\item{bbox}{A bounding box representing the area to be resized. If not
a [terrainr::terrainr_bounding_box] object the function will attempt to
convert it.}

\item{size}{The side length, in pixels, of your eventual .raw tiles.}
}
\value{
A resized [terrainr::terrainr_bounding_box] object.
}
\description{
Resize bounding box to reduce number of blank tiles created
}
\examples{
simulated_data <-  data.frame(id = seq(1, 100, 1),
                              lat = runif(100, 44.04905, 44.17609),
                              lng = runif(100, -74.01188, -73.83493))
bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
make_unity_friendly(bbox)

}
