% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rasters.R
\name{merge_rasters}
\alias{merge_rasters}
\title{Merge multiple raster files into a single raster}
\usage{
merge_rasters(
  input_heightmaps,
  output_heightmap,
  input_orthos = NULL,
  output_ortho = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{input_heightmaps}{A character vector containing the file paths to the
elevation height maps to merge.}

\item{output_heightmap}{The file path to save the merged elevation raster to.
Must be a TIFF file.}

\item{input_orthos}{A character vector, the same length as \code{input_heightmaps},
containing the file paths to the orthoimages to merge. It's assumed that each
orthoimage in the vector corresponds to the height map in the same position
in input_heightmaps. If you don't want to merge any orthoimages, set to NULL.}

\item{output_ortho}{The file path to save the merged orthoimage to.
Must be a TIFF file. Set to NULL if you aren't merging orthoimages.}

\item{overwrite}{Logical: overwrite the output files if they exist?}
}
\description{
Merge multiple raster files into a single raster
}
\examples{
\dontrun{
tile_1 <- terrainr_bounding_box(
  bl = c(44.10379, -74.01177),
  tr = c(44.17573, -73.91171)
)

tile_2 <- terrainr_bounding_box(
  bl = c(44.03184, -74.01177),
  tr = c(44.10379, -73.91171)
)
raster_files <- c(tempfile(), tempfile())
img_bin <- lapply(
  c(tile_1, tile_2),
  function(x) hit_heightmap_api(x, 8000, 8000)
)
writeBin(img_bin[[1]], raster_files[[1]])
writeBin(img_bin[[2]], raster_files[[2]])
merge_rasters(raster_files, tempfile())
}

}
