% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_raw_tiles.R
\name{raster_to_raw_tiles}
\alias{raster_to_raw_tiles}
\title{Crop a raster and convert the output tiles into new formats.}
\usage{
raster_to_raw_tiles(input_file, output_prefix, side_length = 4097, raw = TRUE)
}
\arguments{
\item{input_file}{File path to the input TIFF file to convert.}

\item{output_prefix}{The file path to prefix output tiles with.}

\item{side_length}{The side length, in pixels, for the .raw tiles.}

\item{raw}{Logical: Convert the cropped tiles to .raw? When \code{FALSE}
returns a .png.}
}
\value{
Invisibly, a character vector containing the file paths that were
written to.
}
\description{
This function crops input raster files into smaller square tiles and then
converts them into either .png or .raw files which are ready to be imported
into the Unity game engine.
}
\examples{
\dontrun{
simulated_data <- data.frame(
  id = seq(1, 100, 1),
  lat = runif(100, 44.04905, 44.17609),
  lng = runif(100, -74.01188, -73.83493)
)
bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
bbox <- add_bbox_buffer(bbox, 100)
output_files <- get_tiles(bbox)
temptiff <- tempfile(fileext = ".tif")
merge_rasters(output_files["3DEPElevation"][[1]], temptiff)
raster_to_raw_tiles(temptiff, tempfile())
}

}
\seealso{
Other data manipulation functions: 
\code{\link{georeference_overlay}()},
\code{\link{merge_rasters}()},
\code{\link{vector_to_overlay}()}
}
\concept{data manipulation functions}
