% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiles.R
\name{split_bbox}
\alias{split_bbox}
\title{Split a bounding box into a smaller set of component tiles.}
\usage{
split_bbox(bbox, side_length)
}
\arguments{
\item{bbox}{The bounding box to split into tiles.}

\item{side_length}{The length of each side of the output tiles.}
}
\value{
A list containing the split tiles in position 1, the number of tiles
in the x direction in position 2, and the number of tiles in the y direction
in position 3.
d
}
\description{
This function splits the area contained within a bounding box into a set of
tiles.
}
\keyword{internal}
