% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrainr-package.R
\docType{package}
\name{terrainr-package}
\alias{terrainr}
\alias{terrainr-package}
\title{terrainr: Landscape Visualizations in R and Unity}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Functions for the retrieval, manipulation, and visualization of 
    geospatial data, with an aim towards producing '3D' landscape visualizations
    in the Unity '3D' rendering engine. Functions are also provided for 
    retrieving elevation data and base map tiles from the 'USGS' National Map
    ('<https://viewer.nationalmap.gov/services/>').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikemahoney218.github.io/terrainr/}
  \item \url{https://github.com/mikemahoney218/terrainr}
  \item Report bugs at \url{https://github.com/mikemahoney218/terrainr/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
