% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{terrainr_bounding_box-class}
\alias{terrainr_bounding_box-class}
\title{Bounding boxes in the format expected by [terrainr] functions}
\description{
Bounding boxes in the format expected by [terrainr] functions
}
\section{Slots}{

\describe{
\item{\code{bl}}{A [terrainr::terrainr_coordinate_pair] representing the bottom left
corner of the bounding box}

\item{\code{tr}}{A [terrainr::terrainr_coordinate_pair] representing the top right
corner of the bounding box}
}}

