% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{terrainr_coordinate_pair-class}
\alias{terrainr_coordinate_pair-class}
\title{S4 class for coordinate points in the format expected by [terrainr] functions}
\description{
S4 class for coordinate points in the format expected by [terrainr] functions
}
\section{Slots}{

\describe{
\item{\code{lat}}{Numeric latitude, in decimal degrees}

\item{\code{lng}}{Numeric longitude, in decimal degrees}
}}

\seealso{
Other classes: 
\code{\link{export_bounding_box}()},
\code{\link{export_coord_pair}()},
\code{\link{terrainr_bounding_box-class}},
\code{\link{terrainr_bounding_box}},
\code{\link{terrainr_coordinate_pair}}
}
\concept{classes}
