% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{terrainr_coordinate_point}
\alias{terrainr_coordinate_point}
\title{Construct a terrainr_coordinate_point object}
\usage{
terrainr_coordinate_point(coords, coord_units = c("degrees", "radians"))
}
\arguments{
\item{coords}{A vector of length 2 containing a latitude and longitude. If
unnamed, coordinates are assumed to be in (latitude, longitude) format; if
named, the function will attempt to figure out which value represents which
coordinate.}

\item{coord_units}{String indicating whether coordinates are in degrees or
radians. Degrees stored in radians will be converted to degrees.}
}
\value{
A [terrainr::terrainr_coordinate_point] object
}
\description{
Construct a terrainr_coordinate_point object
}
\examples{
terrainr_coordinate_point(c(44.05003, -74.01164))

}
