% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bbox_buffer.R
\name{add_bbox_buffer}
\alias{add_bbox_buffer}
\title{Add a uniform buffer around a bounding box}
\usage{
add_bbox_buffer(bbox, distance)
}
\arguments{
\item{bbox}{The original bounding box to add a buffer around.}
}
\description{
This function calculates the great circle distance both corners of your
bounding box are from the centroid and extends those by a set distance. Due
to using great circle distance, calculations will not be exact.
}
