% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_haversine_distance.R
\name{calc_haversine_distance}
\alias{calc_haversine_distance}
\title{Extract latitude and longitude from a provided object}
\usage{
calc_haversine_distance(point_1, point_2, coord_units = "degrees")
}
\arguments{
\item{point_1, point_2}{Coordinate pairs (as length-2 numeric vectors with the
names "lat" and "lng") to calculate distance between.}

\item{coord_units}{String indicating whether coordinates are in degrees
(`"degrees"`) or radians (`"radians"`) Degrees stored in radians will be
converted to degrees.}
}
\value{
A vector of length 1 containing distance between points
}
\description{
This is an internal utility function to convert bounding boxes into
coordinate pairs.
}
\seealso{
Other utilities: 
\code{\link{addbuff}},
\code{\link{deg_to_rad}()},
\code{\link{get_centroid}()},
\code{\link{rad_to_deg}()}
}
\concept{utilities}
\keyword{internal}
