% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_distance}
\alias{convert_distance}
\title{Convert distance into meters}
\usage{
convert_distance(
  distance,
  distance_unit = c("meters", "kilometers", "miles", "feet")
)
}
\arguments{
\item{distance}{The numeric distance to be converted to meters}

\item{distance_unit}{A string indicating the units to convert distance from}
}
\description{
Convert distance into meters
}
\examples{
convert_distance(100, "miles")

}
