% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert decimal degrees to radians}
\usage{
deg_to_rad(deg)
}
\arguments{
\item{deg}{A vector of values, in decimal degrees, to convert to radians}
}
\value{
A vector of the same length in radians
}
\description{
Convert decimal degrees to radians
}
\examples{
deg_to_rad(360)
rad_to_deg(deg_to_rad(360))
}
