% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_raw_tiles.R
\name{generate_raw_tiles}
\alias{generate_raw_tiles}
\title{Generate .RAW image tiles of a heightmap}
\usage{
generate_raw_tiles(
  data = NULL,
  lat = NULL,
  lng = NULL,
  bbox = NULL,
  buffer_distance = NULL,
  unity_friendly = NULL,
  tif_filename = NULL,
  raw_output_prefix,
  save_tif = TRUE
)
}
\arguments{
\item{data}{Optionally, a dataframe containing vectors of latitude and
longitude.}

\item{lat}{If \code{data} is not \code{NULL}, the name of the column
containing latitude values. If \code{data} is \code{NULL}, a vector of
latitude values.}

\item{lng}{If \code{data} is not \code{NULL}, the name of the column
containing longitude values. If \code{data} is \code{NULL}, a vector of
longitude values.}

\item{bbox}{Optionally, a bounding box representing the area to pull a DEM
for. If provided, arguments to \code{data}, \code{lat}, and \code{lng} are
ignored.}
}
\description{
This function combines the main functions in the [terrainr] package to create
.RAW image tiles of a USGS digital elevation model (DEM) covering an area of
interest.
}
