% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbox_centroid.R
\name{get_bbox_centroid}
\alias{get_bbox_centroid}
\title{Get the great-circle centroid for a given bounding box.}
\usage{
get_bbox_centroid(bbox)
}
\arguments{
\item{bbox}{The bounding box to find a centroid for. If not already
a \code{\link{terrainr_bounding_box}} object, will be converted.}
}
\value{
A \code{\link{terrainr_coordinate_pair}}.
}
\description{
Get the great-circle centroid for a given bounding box.
}
\examples{
get_bbox_centroid(
  list(
    c(lat = 44.04905, lng = -74.01188),
    c(lat = 44.17609, lng = -73.83493)
  )
)
}
\seealso{
Other utilities: 
\code{\link{addbuff}},
\code{\link{calc_haversine_distance}()},
\code{\link{convert_distance}()},
\code{\link{deg_to_rad}()},
\code{\link{get_coord_bbox}()},
\code{\link{point_from_distance}()},
\code{\link{rad_to_deg}()}
}
\concept{utilities}
\keyword{internal}
