% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbox.R
\name{get_coord_bbox}
\alias{get_coord_bbox}
\title{Get bounding box for set of coordinate points}
\usage{
get_coord_bbox(data = NULL, lat, lng)
}
\arguments{
\item{data}{Optionally, a dataframe containing vectors of latitude and
longitude.}

\item{lat}{If \code{data} is not \code{NULL}, the name of the column
containing latitude values. If \code{data} is \code{NULL}, a vector of
latitude values.}

\item{lng}{If \code{data} is not \code{NULL}, the name of the column
containing longitude values. If \code{data} is \code{NULL}, a vector of
longitude values.}
}
\value{
A list of length 2, containing the bottom-left (named "bl") and
top-right (named "tr") coordinates of the bounding box.
}
\description{
Get bounding box for set of coordinate points
}
\examples{
df <- data.frame(
  lat = c(44.05771, 44.18475),
  lng = c(-73.99212, -73.81515)
)
get_coord_bbox(df, lat, lng)
get_coord_bbox(lat = df$lat, lng = df$lng)
}
