% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_api.R
\name{hit_national_map_api}
\alias{hit_national_map_api}
\title{Hit the USGS 3DEP API and retrieve an elevation heightmap}
\usage{
hit_national_map_api(
  bbox,
  img.width,
  img.height,
  service,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{bbox}{The bounding box (bottom left and top left coordinate pairs)}

\item{img.width}{The number of pixels in the x direction to retrieve}

\item{img.height}{The number of pixels in the y direction to retrieve}

\item{verbose}{Logical: Print out the number of tries required to pull each
tile? Default \code{FALSE}.}
}
\description{
Hit the USGS 3DEP API and retrieve an elevation heightmap
}
\examples{
\dontrun{
hit_national_map_api(
                     bbox = list(c(44.10438, -74.01231),
                                 c(44.17633, -73.91224)),
                     img.width = 8000,
                     img.height = 8000,
                     service = "3DEPElevation")
}

}
