% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rasters.R
\name{merge_rasters}
\alias{merge_rasters}
\title{Merge multiple raster files into a single raster}
\usage{
merge_rasters(input_files, output_file, overwrite = TRUE)
}
\arguments{
\item{input_files}{A character vector containing the file paths to the
rasters to merge.}

\item{output_file}{The file path to save the merged raster file to.}

\item{overwrite}{Logical: overwrite output_file if it exists?}
}
\description{
Merge multiple raster files into a single raster
}
