% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rasters.R
\name{merge_rasters}
\alias{merge_rasters}
\title{Merge multiple raster files into a single raster}
\usage{
merge_rasters(
  input_rasters,
  output_raster = tempfile(fileext = ".tif"),
  input_images = NULL,
  output_image = tempfile(fileext = ".tif"),
  overwrite = TRUE
)
}
\arguments{
\item{input_rasters}{A character vector containing the file paths to the
georeferenced rasters you want to use.}

\item{output_raster}{The file path to save the merged georeferenced raster
to. Must be a TIFF file. Ignored if \code{merge_raster} is not \code{TRUE}.}

\item{input_images}{A character vector, the same length as
\code{input_rasters}, containing the file paths to non-georeferenced images
to merge. It's assumed that each image in the vector corresponds to the
raster in the same position in input_rasters, and these images will be
assigned the same extent and CRS. If you don't want to merge any images,
leave as \code{NULL}.}

\item{output_image}{The file path to save the merged images to.
Must be a TIFF file. Leave as \code{NULL} if you aren't merging images.}

\item{overwrite}{Logical: overwrite the output files if they exist?}
}
\value{
A named list containing the file paths outputs were written to.
}
\description{
Some functions like [get_tiles] return multiple separate files
when it can be useful to have a single larger raster instead. This function
will collapse those multiple raster files into a single TIFF.
}
\examples{
\dontrun{
tiles <- data.frame(
lat = c(44.10379, 44.17573),
lng = c(-74.01177, -73.91171)
)

tiles <- sf::st_as_sf(tiles, coords = c("lng", "lat"))

img_files <- get_tiles(tiles)
merge_rasters(img_files[[1]])

}

}
\seealso{
Other data manipulation functions: 
\code{\link{combine_overlays}()},
\code{\link{georeference_overlay}()},
\code{\link{raster_to_raw_tiles}()},
\code{\link{vector_to_overlay}()}
}
\concept{data manipulation functions}
