% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rasters.R
\name{merge_rasters}
\alias{merge_rasters}
\title{Merge multiple raster files into a single raster}
\usage{
merge_rasters(input_files, output_file, overwrite = TRUE)
}
\arguments{
\item{input_files}{A character vector containing the file paths to the
rasters to merge.}

\item{output_file}{The file path to save the merged raster file to.}

\item{overwrite}{Logical: overwrite output_file if it exists?}
}
\description{
Merge multiple raster files into a single raster
}
\examples{
\dontrun{
tile_1 <- terrainr_bounding_box(
  bl = c(44.10379, -74.01177),
  tr = c(44.17573, -73.91171)
)

tile_2 <- terrainr_bounding_box(
  bl = c(44.03184, -74.01177),
  tr = c(44.10379, -73.91171)
)
raster_files <- c(tempfile(), tempfile())
img_bin <- lapply(
  c(tile_1, tile_2),
  function(x) hit_heightmap_api(x, 8000, 8000)
)
writeBin(img_bin[[1]], raster_files[[1]])
writeBin(img_bin[[2]], raster_files[[2]])
merge_rasters(raster_files, tempfile())
}

}
