% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_rasters.R
\name{merge_rasters}
\alias{merge_rasters}
\title{Merge multiple raster files into a single raster}
\usage{
merge_rasters(
  input_rasters,
  output_raster = NULL,
  input_images = NULL,
  output_image = NULL,
  overwrite = TRUE,
  merge_raster = TRUE
)
}
\arguments{
\item{input_rasters}{A character vector containing the file paths to the
georeferenced rasters you want to use.}

\item{output_raster}{The file path to save the merged georeferenced raster to.
Must be a TIFF file. Ignored if \code{merge_raster} is not \code{TRUE}.}

\item{input_images}{A character vector, the same length as \code{input_rasters},
containing the file paths to non-georeferenced images to merge. It's assumed
that each image in the vector corresponds to the raster in the same position
in input_rasters, and these images will be assigned the same extent and CRS.
If you don't want to merge any images, leave as \code{NULL}.}

\item{output_image}{The file path to save the merged images to.
Must be a TIFF file. Leave as \code{NULL} if you aren't merging images.}

\item{overwrite}{Logical: overwrite the output files if they exist?}

\item{merge_raster}{Logical: Should the files provided to input_rasters be
merged? Set to \code{FALSE} if you're only providing these input tiles to
georeference tiles in \code{input_images}.}
}
\value{
A named list of length 2 with the original inputs to output_raster
and output_image
}
\description{
Some functions like \code{\link{get_tiles}} return multiple separate files
when it can be useful to have a single larger raster instead. This function
will collapse those multiple raster files into a single TIFF.
}
\details{
Additionally, some outputs from \code{\link{get_tiles}} (such as when
\code{service = "ortho"}) are not georeferenced, making future processing
with them harder. This function can georeference those images using a
reference raster, then merge them into a single object for further processing.
}
\examples{
\dontrun{
tile_1 <- terrainr_bounding_box(
  bl = c(lat = 44.10379, lng = -74.01177),
  tr = c(lat = 44.17573, lng = -73.91171)
)

tile_2 <- terrainr_bounding_box(
  bl = c(lat = 44.03184, lng = -74.01177),
  tr = c(lat = 44.10379, lng = -73.91171)
)
raster_files <- c(tempfile(), tempfile())
img_bin <- lapply(
  c(tile_1, tile_2),
  function(x) hit_national_map_api(x, 8000, 8000, "3DEPElevation")
)
writeBin(img_bin[[1]], raster_files[[1]])
writeBin(img_bin[[2]], raster_files[[2]])
merge_rasters(raster_files, tempfile())
}

}
\seealso{
Other data manipulation functions: 
\code{\link{raster_to_raw_tiles}()}
}
\concept{data manipulation functions}
