% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{Convert radians to degrees}
\usage{
rad_to_deg(rad)
}
\arguments{
\item{rad}{A vector of values, in radians, to convert to decimal degrees}
}
\value{
A vector of the same length in decimal degrees
}
\description{
Convert radians to degrees
}
\examples{
rad_to_deg(2 * base::pi)
rad_to_deg(deg_to_rad(360))
}
