% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_to_raw_tiles.R
\name{raster_to_raw_tiles}
\alias{raster_to_raw_tiles}
\title{Convert raster file into .raw image tiles}
\usage{
raster_to_raw_tiles(input_file, output_prefix, side_length = 4097)
}
\arguments{
\item{input_file}{File path to the input GeoTIFF file to convert.}

\item{output_prefix}{The file path to prefix output tiles with.}

\item{side_length}{The side length, in pixels, for the .raw tiles.}
}
\description{
This function converts raster objects (stored as GeoTIFF files) into .raw
image square tiles.
}
\examples{
\dontrun{
simulated_data <- data.frame(
  id = seq(1, 100, 1),
  lat = runif(100, 44.04905, 44.17609),
  lng = runif(100, -74.01188, -73.83493)
)

bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
bbox <- add_bbox_buffer(bbox, 100)
bbox <- make_unity_friendly(bbox)
temptiff <- tempfile(fileext = ".tif")
get_heightmap_tiles(temptiff)
raster_to_raw_tiles(temptiff, tempfile())
}

}
