% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiles.R
\name{split_bbox}
\alias{split_bbox}
\title{Split a bounding box into a smaller set of component tiles.}
\usage{
split_bbox(bbox, side_length)
}
\arguments{
\item{bbox}{The bounding box to split into tiles.}

\item{side_length}{The length of each side of the output tiles.}
}
\value{
A list containing the split tiles in position 1, the number of tiles
in the x direction in position 2, and the number of tiles in the y direction
in position 3.
}
\description{
This function splits the area contained within a bounding box into a set of
tiles.
}
\examples{
simulated_data <- data.frame(
  id = seq(1, 100, 1),
  lat = runif(100, 44.04905, 44.17609),
  lng = runif(100, -74.01188, -73.83493)
)

bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
bbox <- add_bbox_buffer(bbox, 100)
split_bbox(bbox, 4096)

}
