% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrainr-package.R
\docType{package}
\name{terrainr-package}
\alias{terrainr}
\alias{terrainr-package}
\title{terrainr: Retrieve and Transform 'USGS' Digital Elevation Models}
\description{
Functions for the retrieval and manipulation of 'USGS' digital 
    elevation models, making it much easier for users to produce tiled images of 
    areas of interest which can then be imported into 3D rendering programs such 
    as Unity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikemahoney218/terrainr}
  \item Report bugs at \url{https://github.com/mikemahoney218/terrainr/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
