% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrainr-package.R
\docType{package}
\name{terrainr-package}
\alias{terrainr}
\alias{terrainr-package}
\title{terrainr: Retrieve and Transform Data From The 'USGS' National Map}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Functions for the retrieval and manipulation of 'USGS' National Map 
    data, enabling users to download elevation and image data for areas of 
    interest. Functions are also provided to transform these data sources into 
    formats that can be used to create 3D elevation models in the Unity 
    rendering engine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikemahoney218/terrainr}
  \item Report bugs at \url{https://github.com/mikemahoney218/terrainr/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
