% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrainr-package.R
\docType{package}
\name{terrainr-package}
\alias{terrainr}
\alias{terrainr-package}
\title{terrainr: Landscape Visualizations in R and 'Unity'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Functions for the retrieval, manipulation, and visualization of 
    'geospatial' data, with an aim towards producing '3D' landscape 
    visualizations in the 'Unity' '3D' rendering engine. Functions are also 
    provided for retrieving elevation data and base map tiles from the 'USGS' 
    National Map ('<https://apps.nationalmap.gov/services/>').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/terrainr/}
  \item \url{https://github.com/ropensci/terrainr}
  \item Report bugs at \url{https://github.com/ropensci/terrainr/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

Other contributors:
\itemize{
  \item Mike Johnson (Mike reviewed the package (v. 0.2.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/416>) [reviewer]
  \item Sydney Foks (Sydney reviewed the package (v. 0.2.1) for rOpenSci, see <https://github.com/ropensci/software-review/issues/416>) [reviewer]
}

}
\keyword{internal}
