% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{terrainr_bounding_box-class}
\alias{terrainr_bounding_box-class}
\title{S4 class for bounding boxes in the format expected by \code{terrainr}
functions.}
\description{
S4 class for bounding boxes in the format expected by \code{terrainr}
functions.
}
\section{Slots}{

\describe{
\item{\code{bl}}{A \code{\link{terrainr_coordinate_pair}} representing the bottom
left corner of the bounding box}

\item{\code{tr}}{A \code{\link{terrainr_coordinate_pair}} representing the top right
corner of the bounding box}
}}

\keyword{internal}
