% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{terrainr_coordinate_pair-class}
\alias{terrainr_coordinate_pair-class}
\title{Coordinate points in the format expected by [terrainr] functions}
\description{
Coordinate points in the format expected by [terrainr] functions
}
\section{Slots}{

\describe{
\item{\code{lat}}{Numeric latitude, in decimal degrees}

\item{\code{lng}}{Numeric longitude, in decimal degrees}
}}

