% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{terrainr_coordinate_pair-class}
\alias{terrainr_coordinate_pair-class}
\title{S4 class for coordinate points in the format expected by
\code{terrainr} functions.}
\description{
S4 class for coordinate points in the format expected by
\code{terrainr} functions.
}
\section{Slots}{

\describe{
\item{\code{lat}}{Numeric latitude, in decimal degrees}

\item{\code{lng}}{Numeric longitude, in decimal degrees}
}}

\seealso{
Other classes and related functions: 
\code{\link{terrainr_coordinate_pair}}
}
\concept{classes and related functions}
\keyword{internal}
