% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{terrainr_coordinate_pair}
\alias{terrainr_coordinate_pair}
\title{Construct a terrainr_coordinate_pair object.}
\usage{
terrainr_coordinate_pair(coords, coord_units = c("degrees", "radians"))
}
\arguments{
\item{coords}{A vector of length 2 containing a latitude and longitude. If
unnamed, coordinates are assumed to be in (latitude, longitude) format; if
named, the function will attempt to figure out which value represents which
coordinate. Currently this function understands "lat", "latitude", and "y" as
names for latitude and "lng", "long", "longitude", and "x" for longitude.}

\item{coord_units}{String indicating whether coordinates are in degrees or
radians. Degrees stored in radians will be converted to degrees.}
}
\value{
\code{terrainr_coordinate_pair} object
}
\description{
In order to simplify code, most \code{terrainr} functions expect a set S4
class representation of coordinate pairs and bounding boxes. If the provided
data isn't in the expected S4 format, these functions are used to cast the
data into the target class.
}
\seealso{
Other classes and related functions: 
\code{\link{terrainr_coordinate_pair-class}}
}
\concept{classes and related functions}
\keyword{internal}
